+++ app.js
@@ -28,9 +28,8 @@ var cookieParser = require("cookie-parser");
 var bodyParser = require("body-parser");
 
 var routes = require("./routes");
-var photos = require("./routes/photo")(db);
-var selfie = require("./routes/selfie");
-var monitor = require("./routes/monitor");
+var selfie = require("./routes/selfie")(db);
+var meme = require("./routes/meme");
 
 var app = express();
 app.use(auth.connect(basic));
@@ -51,15 +50,20 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 app.use(app.router);
 
-app.get("/", routes.index);
+//Take selfie
+app.get("/selfie", selfie.showselfie);
+app.get("/selfies", selfie.showphotos);
+app.post("/takeselfieshot", selfie.takesnapshot);
 
-app.get("/photos", photos.showphotos);
-app.get("/photosjson", photos.photosjson);
-app.get("/photo/:photoid", photos.photo);
-app.post("/takesnapshot", photos.takesnapshot);
-app.post("/takememeshot", selfie.takememeshot(db));
-app.get("/filter", selfie.showfilter);
-app.get("/memes", selfie.showmemes(db));
+//Take meme
+app.get("/meme", meme.showfilter);
+app.get("/memes", meme.showmemes(db));
+app.post("/takememeshot", meme.takememeshot(db));
+
+//General paths
+app.get("/", routes.index);
+app.get("/photosjson", selfie.photosjson);
+app.get("/photo/:photoid", selfie.photo);
 
 /// catch 404 and forwarding to error handler
 app.use(function(req, res, next) {
