+++ public/javascripts/reloader.js
@@ -5,7 +5,7 @@
     setInterval(doTheCheck, 120000); // Every two minutes
 
     function doTheCheck() {
-        $.get('/monitor').then(function(response) {
+        $.get('/').then(function(response) {
             if (serverNotResponding) {
                 window.location.reload(true);
                 serverNotResponding = false;
