+++ routes/index.js
@@ -1,6 +1,5 @@
 "use strict";
 
-/* GET home page. */
 exports.index = function(req, res){
-  res.render("index", { title: "Snapshot" });
+  res.json({ statusCode: 200 });
 };
