--- routes/monitor.js
@@ -1,5 +1,0 @@
-"use strict";
-
-exports.index = function(req, res){
-  res.json({ statusCode: 200 });
-};
