+++ routes/selfie.js
@@ -10,6 +10,7 @@ var _ = require("lodash");
 var knox = require("knox");
 var env = require("../env");
 
+// Knox is used to send images to S3
 var client = knox.createClient({
   key: env.KNOX_KEY,
   secret: env.KNOX_SECRET,
@@ -17,30 +18,65 @@ var client = knox.createClient({
   region: env.KNOX_REGION
 });
 
-// Few safety precaucions
-assert(fs.statSync(PHOTOSDATADIR).isDirectory());
+module.exports = function(db) {
+  var exports = { };
 
-var imageURI;
+  // Few safety precaucions
+  assert(fs.statSync(PHOTOSDATADIR).isDirectory());
 
-exports.showmemes = function(db){
-  return function(req, res) {
+  exports.showselfie = function(req, res){
+    res.render("selfie", { title: "Snapshot" });
+  };
+
+  exports.showphotos = function(req, res) {
     var collection = db.get("photocollection");
-    collection.find({location: "M"}, function(e, docs){
-      _(docs).reverse();
-      res.render("memes", { "images": _.take(docs, 20)});
+    collection.find({location: "S"}, function(e, docs){
+      if(docs) {
+        _(docs).reverse();
+        res.render("photos", { "images": _.take(docs, 20)});
+      } else {
+        res.render("photos", { "images": []});
+      }
     });
   };
-};
 
-exports.takememeshot = function(db) {
-  return function(req, res) {
-    var reqId = _.uniqueId("meme");
+  exports.photosjson = function(req, res) {
+    var collection = db.get("photocollection");
+    collection.find({}, {}, function(e, docs){
+      var data = {};
+      if(docs) {
+        _(docs).reverse();
+        data = JSON.stringify(docs);
+        res.writeHead(200, {
+          "Content-Length": data.length,
+          "Content-Type": "application/json"
+        });
+      }
+      res.send(data);
+    });
+  };
 
+  exports.filterphotos = function(req, res) {
+    res.render("filter");
+  };
+
+  exports.photo = function (req, res) {
+    var photoId = req.params.photoid || "foobar";
+    var filename = photoId + ".png";
+    client.getFile(filename, function(err, stream) {
+      if(!err) {
+        stream.pipe(res);
+      } else {
+        console.log("oh snap, error ocurred");
+      }
+    });
+  };
+
+  exports.takesnapshot = function(req, res) {
+    var reqId = _.uniqueId("snap");
+    console.log(reqId, "Entered takesnapshot");
     try {
-
       var string = req.body.data;
-      //Maybe not the right way to do this?
-      imageURI = string;
       var regex = /^data:.+\/(.+);base64,(.*)$/;
       var matches = string.match(regex);
       var ext = matches[1];
@@ -48,92 +84,75 @@ exports.takememeshot = function(db) {
       var buffer = new Buffer(data, "base64");
       var now = new XDate();
       var timeAsISO = now.toISOString().replace(/[:\-]/g, "");
-      var id = "M" + timeAsISO + "-" + _.uniqueId();
+      var id = "S" + timeAsISO + "-" + _.uniqueId();
       var fileName = id + "." + ext;
+      console.log(reqId, "file name", fileName);
+      console.log(reqId, "data length", data.length);
 
-      // save file to S3
-      saveToS3(reqId, buffer, fileName)
-        .then(function () {
-          return postToTwitter(reqId, buffer);
-        })
-        .then(function (twitterUrl) {
-          return saveToDB(reqId, db, id, timeAsISO, ext, twitterUrl);
-        })
-        .catch(function (err) {
-          console.error(reqId, "ERROR", err);
-          res.status(500).json({
-            error: "" + err
-          });
-        })
-        .done(function () {
-          res.json({});
-        });
-    } catch (err) {
-      console.error("ERROR", err);
-      res.json(500).json({
-        error: "" + err
+      saveImageToS3(reqId, buffer, fileName)
+      .then(function () {
+        return postToTwitter(reqId, buffer, now);
+      })
+      .then(function (urlToTwitter) {
+        console.log(reqId, "twitter responded with url", urlToTwitter);
+        return saveToDB(reqId, db, id, timeAsISO, ext, urlToTwitter);
+      })
+      .catch(function (error) {
+        console.error(reqId, "Saving failed big time", error);
+        res.status(500).json({ error: error });
+      })
+      .done(function(){
+        res.json({ });
       });
+    } catch (err) {
+      console.error(reqId, "ERROR", err);
+      res.status(500).json({ error: "" + err });
     }
   };
+  return exports;
 };
 
-exports.showfilter = function(req, res) {
-  res.render("filter", { title: "Filter", img: imageURI });
-};
-
-function saveToS3(reqId, buffer, fileName) {
+function saveImageToS3(reqId, buffer, fileName) {
   var deferred = when.defer();
-
   var headers = {"Content-Type": "text/plain"};
   client.putBuffer(buffer, "/" + fileName, headers, function(err, res) {
-       var status = res.statusCode;
-
-       console.log("in put: " + status);
-
-       if (err) {
-         console.error(reqId, "Error saving to S3", err);
-         deferred.reject(err);
-         return;
-       }
-
-       if (200 !== res.statusCode) {
-      deferred.reject(new Error("Status code 200 != " + status));
+    console.log(reqId, "putBuffer, err - res " + err + " - " + res);
+    if(res && res.statusCode === 200) {
+      deferred.resolve("photo saved to S3");
+    } else {
+      console.error(reqId, "ERR failed to put to buffer");
+      deferred.reject(new Error("Save to S3 failed"));
     }
-
-    deferred.resolve("saved");
   });
 
   return deferred.promise;
 }
 
-function postToTwitter(reqId, buffer) {
+function postToTwitter(reqId, buffer, now) {
   var deferred = when.defer();
-  var now = new XDate();
-  console.log("tweet");
+
   var r = request.post({
     url: "https://api.twitter.com/1.1/statuses/update_with_media.json",
-    // snapshotNowMeme
     oauth: {
       "consumer_key": env.TWITTER_CONSUMER_KEY,
       "consumer_secret": env.TWITTER_CONSUMER_SECRET,
       "token": env.TWITTER_TOKEN,
       "token_secret": env.TWITTER_TOKEN_SECRET
     }
-  }, function(err, response, body) {
-      if (!err && response && response.statusCode === 200) {
-          console.log("tweeted");
+  }, function (err, response, body) {
+    console.log(reqId, "posted tweet: err - response " + err + " - " + response);
+    if (response) {
       deferred.resolve(body);
-      } else {
-          console.log("tweet failed, err " + err);
-          if (env.TWITTER_CONSUMER_KEY === "") {
+    } else {
+      if (env.TWITTER_CONSUMER_KEY === "") {
         deferred.resolve(body);
-          } else {
-        deferred.reject(err);
+      } else {
+        deferred.reject(new Error("Post to Twitter failed"));
       }
     }
   });
 
-  var status = "New Meme: " + now.toString("d.M.yyyy HH:mm:ss");
+  var status = "New Visitor: " + now.toString("d.M.yyyy HH:mm:ss");
   var form = r.form();
   form.append("status", status);
   form.append("media[]", buffer);
@@ -156,7 +175,7 @@ function saveToDB(reqId, db, id, timeAsISO, ext, twitterUrl) {
   var collection = db.get("photocollection");
   collection.insert({
     id: id,
-    location: "M",
+    location: "S",
     url: "/photo/" + id,
     twitter_url: twitterUrl,
     time: timeAsISO,
