--- views/monitor.hjs
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Monitor</title>
-  </head>
-  <body>
-    ok
-  </body>
-</html>
